% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday Problem Probability}
\usage{
birthday(x)
}
\arguments{
\item{x}{Integer vector. The sample size(s), i.e., the number of people in the group.}
}
\value{
A numeric vector of probabilities, corresponding to each input value of `x`.
}
\description{
Computes the probability that two or more people in a group of size `x`
share the same birthday, assuming 365 equally likely birthdays and no leap years.
}
\examples{
birthday(20)
birthday(20:25)
}
