% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firedocumentation.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire dataset}
\format{
A data frame with X rows and Y variables.
}
\source{
Example or simulated data (replace this with real source if known).
}
\usage{
fire
}
\description{
A sample dataset included in the MATH4753F25dastin package.
}
\keyword{datasets}
