% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt_mean.R
\docType{package}
\name{MATH4753F25dastin-package}
\alias{MATH4753F25dastin}
\alias{MATH4753F25dastin-package}
\title{#' @keywords internal}
\description{
Lab code for 2025.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dastin-tain/MATH4753F25dastin}
  \item Report bugs at \url{https://github.com/dastin-tain/MATH4753F25dastin/issues}
}

}
\author{
\strong{Maintainer}: Dastin Scott \email{dastinscott@gmail.com}

}
