% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Simulation Function}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{Number of iterations (simulated experiments).}

\item{n}{Number of trials per experiment.}

\item{p}{Probability of success on each trial.}
}
\value{
A table of relative frequencies of the number of successes across iterations.
}
\description{
Simulates repeated binomial experiments and produces a barplot of the distribution of successes.
}
\examples{
mybin(iter=1000, n=10, p=0.7)
}
