% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Confidence Interval}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap samples.}

\item{x}{Numeric sample data.}

\item{fun}{Function name (as character) for the statistic.}

\item{alpha}{Significance level for CI.}

\item{cx}{Text size for labels.}

\item{...}{Extra graphical options.}
}
\value{
List with bootstrap CI, statistic name, data, and resampled values.
}
\description{
Bootstrap Confidence Interval
}
\examples{
set.seed(39)
x <- rnorm(25, 25, 10)
myboot2(10000, x, fun = "mean", alpha = 0.05, col = "purple")
}
