% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum likelihood by grid search}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Log-likelihood function of the form lfun(x, param).}

\item{x}{Data vector.}

\item{param}{Grid of parameter values.}

\item{...}{Additional graphical arguments passed to plot.}
}
\value{
A list with the MLE and log-likelihood value.
}
\description{
Maximum likelihood by grid search
}
