% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the normal distribution.}

\item{a}{The value of x up to which the area is shaded.}
}
\value{
A list with mu, sigma, and the probability P(X <= a).
}
\description{
Plots a normal curve and shades the area from -Inf to x = a.
}
\examples{
myncurve(mu = 10, sigma = 5, a = 6)
}
