% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets - Overbooking optimization}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats for cheeks available on the flight.}

\item{gamma}{Probability that the flight will be overbooked}

\item{p}{Probability that a passenger shows up.}
}
\value{
list:
\itemize{
  \item \code{nd} - tickets for discrete cheeks
  \item \code{nc} - tickets for continuous cheeks
  \item \code{N} - number of seats for cheeks
  \item \code{p} - decent human probability
  \item \code{gamma} - PAIN probability
  }
}
\description{
Calculates the optimal number of tickets to sell for a flight with N seats,
given a show-up probability and an acceptable overbooking probability.
Produces discrete (Binomial) and continuous (Normal approximation) results and plots.
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
